Move cryptokey service to separate service part
